Renamed constant, added image to content
