+++ src/components/blogs/blogContent/blogContent.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Post } from '@/types/blogs';
+import { Post } from '@/types/posts';
 import BlogService from '@/services/blog.service';
 import Image from 'next/image';
 import Link from 'next/link';
