+++ src/components/blogs/recentBlogCard/recentBlogCard.tsx
@@ -1,4 +1,4 @@
-import { Post } from '@/types/blogs';
+import { Post } from '@/types/posts';
 import Image from 'next/image';
 import React from 'react';
 import Link from 'next/link';
