+++ src/components/home/homeBanner/homeBanner.tsx
@@ -17,7 +17,7 @@ const HomeBanner = async () => {
         <Link href="/about">
           <div className={styles.bannerWelcomeImage}>
             <Image
-              src="https://static.wixstatic.com/media/2e2a49_d456d06d41b346d3b36f426cb4142859~mv2.jpg/v1/crop/x_1646,y_0,w_3950,h_3840/fill/w_144,h_140,al_c,q_80,usm_0.66_1.00_0.01,enc_auto/GettyImages-1145254743.jpg"
+              src={content.banner.imageUrl}
               alt="Blog"
               width={144}
               height={144}
