+++ src/components/projects/projectCard/projectCard.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import Link from 'next/link';
 import React from 'react';
-import { Post } from '@/types/blogs';
+import { Post } from '@/types/posts';
 import styles from './projectCard.module.css';
 
 type Props = {
